+++ core/src/jsMain/kotlin/InitNavigation.kt
@@ -26,6 +26,7 @@ inline fun <reified Base : NavigationNodeDefaultConfig> initNavigation(
     scope: CoroutineScope = CoroutineScope(Dispatchers.Main),
     savingDebounce: Long = 0L,
     rootChain: NavigationChain<Base>,
+    dropRedundantChainsOnRestore: Boolean = false,
     nodesFactory: NavigationNodeFactory<Base>
 ): Job {
     val logger = TagLogger("NavigationJob")
@@ -61,7 +62,8 @@ inline fun <reified Base : NavigationNodeDefaultConfig> initNavigation(
         restoreHierarchy<Base>(
             existsChain ?: startChain,
             factory = resultNodesFactory,
-            rootChain = rootChain
+            rootChain = rootChain,
+            dropRedundantChainsOnRestore = dropRedundantChainsOnRestore
         ) ?.start(subscope)
     }
 }
@@ -76,6 +78,7 @@ inline fun <reified Base : NavigationNodeDefaultConfig> initNavigation(
     scope: CoroutineScope = CoroutineScope(Dispatchers.Main),
     savingDebounce: Long = 0L,
     rootChain: NavigationChain<Base>? = null,
+    dropRedundantChainsOnRestore: Boolean = false,
     nodesFactory: NavigationNodeFactory<Base>
 ): Job = initNavigation(
     startChain,
@@ -83,5 +86,6 @@ inline fun <reified Base : NavigationNodeDefaultConfig> initNavigation(
     scope,
     savingDebounce,
     rootChain ?: NavigationChain<Base>(null, nodesFactory),
+    dropRedundantChainsOnRestore,
     nodesFactory
 )