+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `NavigationChain` got its own optional id
 * `NavigationChain` and `NavigationNode` got `findChain` extension
 * `NavigationChain` and `NavigationNode` got `findNode` extension
+* On restoring of hierarchy it is possible to remove the chains which do not fit restoring one
 
 ## 0.0.15